Refactor(back): ERD 수정에 따른 User 도메인 재정의

See merge request s14-webmobile4-sub1/S14P11D102!15

+++ Backend/src/main/java/com/ssafy/closetory/config/security/JwtTokenProvider.java
@@ -23,7 +23,7 @@ public class JwtTokenProvider {
   }
 
   //  Access Token 생성
-  public String createAccessToken(Long userId) {
+  public String createAccessToken(Integer userId) {
     Date now = new Date();
     Date expiry = new Date(now.getTime() + accessTokenValidityTime);
